requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
twilio==8.10.3
razorpay==1.3.0
python-dotenv==1.0.0
pillow>=9.0.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
jinja2==3.1.2
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4
